backend/ml-service/requirements.txt
# Core ML and API dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# ML Libraries - PRODUCTION OPTIMIZED
scikit-learn==1.3.2
pyod==1.1.3
# xgboost==2.0.2  # Removed - not used in hot path
# tensorflow==2.15.0  # Removed - 2GB+ bloat, not used in production
# keras==2.15.0  # Removed - not used in production

# Data Processing
pandas==2.1.4
numpy==1.24.4
# polars==0.20.2  # Removed - not yet integrated

# Model Management
mlflow==2.8.1
joblib==1.3.2

# Feature Engineering
networkx==3.2.1
scipy==1.11.4

# Monitoring and Logging
prometheus-client==0.19.0
structlog==23.2.0
psutil==5.9.6

# Async and HTTP
httpx==0.25.2
aiofiles==23.2.0

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
flake8==6.1.0

# Configuration
pyyaml==6.0.1
python-dotenv==1.0.0

# Security
cryptography==41.0.8

# Production Performance
gunicorn==21.2.0  # Multi-worker WSGI server
